filename secured, added custom filter class
