+++ forms-flow-documents/src/formsflow_documents/helpers/__init__.py
@@ -1,3 +1,4 @@
 """PDF export helper modules."""
 
+from .custom_filters import TemplateFilters
 from .helper import PdfHelpers
